requirements.txt
Pillow>=9.0.0
playsound>=1.3.0